Abstract TreeInpsector into Label/List/Menu
